+++ geotrek/altimetry/tests/test_elevation.py
@@ -4,12 +4,16 @@ from django.db import connections, DEFAULT_DB_ALIAS
 from django.contrib.gis.geos import MultiLineString, LineString
 from django.core.management import call_command
 from django.core.management.base import CommandError
+from django.test.utils import override_settings
 
 from geotrek.core.models import Path
 from geotrek.core.factories import TopologyFactory
 from geotrek.altimetry.helpers import AltimetryHelper
 
 import os
+import sys
+import mock
+from StringIO import StringIO
 
 
 class ElevationTest(TestCase):
@@ -306,10 +310,23 @@ class CommandLoadDemTest(TestCase):
     """
     TODO: We will need to replace raster management with geodjango (>= py3 / django 2.0).
     """
+    def test_fail_import(self):
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with mock.patch.dict(sys.modules, {'osgeo': None}):
+            with self.assertRaises(CommandError) as e:
+                call_command('loaddem', filename, '--replace', verbosity=0)
+            self.assertEqual('GDAL Python bindings are not available. Can not proceed.', e.exception.message)
 
     def test_success(self):
+        output_stdout = StringIO()
+        conn = connections[DEFAULT_DB_ALIAS]
+        cur = conn.cursor()
+        cur.execute('CREATE TABLE mnt (rid serial primary key, rast raster)')
         filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
-        call_command('loaddem', filename, '--replace', verbosity=0)
+        call_command('loaddem', filename, '--replace', verbosity=2, stdout=output_stdout)
+        self.assertIn('DEM successfully loaded.', output_stdout.getvalue())
+        self.assertIn('DEM successfully clipped/projected.', output_stdout.getvalue())
+        self.assertIn('Everything looks fine, we can start loading DEM', output_stdout.getvalue())
         conn = connections[DEFAULT_DB_ALIAS]
         cur = conn.cursor()
         cur.execute('SELECT ST_Value(rast, ST_SetSRID(ST_MakePoint(602500, 6650000), 2154)) FROM mnt;')
@@ -340,3 +357,60 @@ class CommandLoadDemTest(TestCase):
         with self.assertRaises(CommandError) as e:
             call_command('loaddem', filename, verbosity=0)
         self.assertIn('DEM format is not recognized by GDAL.', e.exception)
+
+    @override_settings(SPATIAL_EXTENT=(0, 0, 0, 0))
+    def test_bbox_not_intersect(self):
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with self.assertRaises(CommandError) as e:
+            call_command('loaddem', filename, '--replace', verbosity=0)
+        self.assertIn('DEM file does not match project extent', e.exception.message)
+
+    @mock.patch('geotrek.altimetry.management.commands.loaddem.Command.call_command_system')
+    def test_fail_raster2pgsql_first(self, sp):
+        def command_fail_raster(cmd, **kwargs):
+            if 'raster2pgsql -G > /dev/null' in cmd:
+                return 1
+            return 0
+        sp.side_effect = command_fail_raster
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with self.assertRaises(CommandError) as e:
+            call_command('loaddem', filename, '--replace', verbosity=0)
+        self.assertEqual('Caught Exception: raster2pgsql failed with exit code 1', e.exception.message)
+
+    @mock.patch('geotrek.altimetry.management.commands.loaddem.Command.call_command_system')
+    def test_fail_gdalwarp(self, sp):
+        def command_fail_gdalwarp(cmd, **kwargs):
+            if 'gdalwarp' in cmd:
+                return 1
+            return 0
+        sp.side_effect = command_fail_gdalwarp
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with self.assertRaises(CommandError) as e:
+            call_command('loaddem', filename, '--replace', verbosity=0)
+        self.assertEqual('Caught Exception: gdalwarp failed with exit code 1', e.exception.message)
+
+    @mock.patch('geotrek.altimetry.management.commands.loaddem.Command.call_command_system')
+    def test_fail_raster2pgsql_second(self, sp):
+        def command_fail_raster(cmd, **kwargs):
+            if 'raster2pgsql -c -C -I -M -t' in cmd:
+                return 1
+            return 0
+        sp.side_effect = command_fail_raster
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with self.assertRaises(CommandError) as e:
+            call_command('loaddem', filename, '--replace', verbosity=0)
+        self.assertEqual('Caught Exception: raster2pgsql failed with exit code 1', e.exception.message)
+
+    @mock.patch('osgeo.gdal.Dataset.GetProjection', return_value='')
+    def test_fail_projection(self, sp):
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with self.assertRaises(CommandError) as e:
+            call_command('loaddem', filename, '--replace', verbosity=0)
+        self.assertEqual('DEM coordinate system is unknown.', e.exception.message)
+
+    @mock.patch('osgeo.gdal.Dataset.GetGeoTransform', return_value=None)
+    def test_fail_extent(self, sp):
+        filename = os.path.join(os.path.dirname(__file__), 'data', 'elevation.tif')
+        with self.assertRaises(CommandError) as e:
+            call_command('loaddem', filename, '--replace', verbosity=0)
+        self.assertEqual('DEM extent is unknown.', e.exception.message)