+++ geotrek/infrastructure/factories.py
@@ -15,6 +15,14 @@ class InfrastructureTypeFactory(factory.DjangoModelFactory):
     pictogram = dummy_filefield_as_sequence('thumbnail %s')
 
 
+class InfrastructureTypeNoPictogramFactory(factory.DjangoModelFactory):
+    class Meta:
+        model = models.InfrastructureType
+
+    label = factory.Sequence(lambda n: u"Type %s" % n)
+    type = models.INFRASTRUCTURE_TYPES.BUILDING
+
+
 class InfrastructureConditionFactory(factory.DjangoModelFactory):
     class Meta:
         model = models.InfrastructureCondition
@@ -31,11 +39,28 @@ class InfrastructureFactory(TopologyFactory):
     published = True
 
 
+class InfrastructureNoPictogramFactory(TopologyFactory):
+    class Meta:
+        model = models.Infrastructure
+    name = factory.Sequence(lambda n: u"Infrastructure %s" % n)
+    type = factory.SubFactory(InfrastructureTypeNoPictogramFactory)
+    condition = factory.SubFactory(InfrastructureConditionFactory)
+    published = True
+
+
 class SignageFactory(TopologyFactory):
     class Meta:
         model = models.Signage
     name = factory.Sequence(lambda n: u"Signage %s" % n)
     type = factory.SubFactory(InfrastructureTypeFactory, type=models.INFRASTRUCTURE_TYPES.SIGNAGE)
     condition = factory.SubFactory(InfrastructureConditionFactory)
-    factory.SubFactory(InfrastructureTypeFactory)
+    published = True
+
+
+class SignageNoPictogramFactory(TopologyFactory):
+    class Meta:
+        model = models.Signage
+    name = factory.Sequence(lambda n: u"Signage %s" % n)
+    type = factory.SubFactory(InfrastructureTypeNoPictogramFactory, type=models.INFRASTRUCTURE_TYPES.SIGNAGE)
+    condition = factory.SubFactory(InfrastructureConditionFactory)
     published = True