+++ geotrek/infrastructure/models.py
@@ -10,7 +10,7 @@ from mapentity.models import MapEntityMixin
 
 from geotrek.common.utils import classproperty
 from geotrek.core.models import Topology, Path
-from geotrek.authent.models import StructureRelatedManager, StructureRelated, StructureOrNoneRelated
+from geotrek.authent.models import StructureRelated, StructureOrNoneRelated
 from geotrek.common.mixins import BasePublishableMixin, OptionalPictogramMixin
 
 
@@ -82,7 +82,7 @@ class InfrastructureCondition(StructureOrNoneRelated):
         return self.label
 
 
-class BaseInfrastructure(MapEntityMixin, BasePublishableMixin, Topology, StructureRelated):
+class BaseInfrastructure(BasePublishableMixin, Topology, StructureRelated):
     """ A generic infrastructure in the park """
     topo_object = models.OneToOneField(Topology, parent_link=True,
                                        db_column='evenement')
@@ -101,6 +101,7 @@ class BaseInfrastructure(MapEntityMixin, BasePublishableMixin, Topology, Structu
                            db_column='id_externe')
 
     class Meta:
+        abstract = True
         db_table = 'a_t_amenagement'
 
     def __unicode__(self):
@@ -129,9 +130,7 @@ class BaseInfrastructure(MapEntityMixin, BasePublishableMixin, Topology, Structu
 
     @property
     def cities_display(self):
-        if hasattr(self, 'cities'):
-            return [unicode(c) for c in self.cities]
-        return []
+        return [unicode(c) for c in self.cities] if hasattr(self, 'cities') else []
 
     @classproperty
     def cities_verbose_name(cls):
@@ -149,19 +148,12 @@ class InfrastructureGISManager(gismodels.GeoManager):
         return all_years
 
 
-class InfrastructureStructureManager(StructureRelatedManager):
-    """ Overide default structure related manager, and filter by type. """
-    def get_queryset(self):
-        return super(InfrastructureStructureManager, self).get_queryset().exclude(type__type=INFRASTRUCTURE_TYPES.SIGNAGE)
-
-
-class Infrastructure(BaseInfrastructure):
+class Infrastructure(MapEntityMixin, BaseInfrastructure):
     """ An infrastructure in the park, which is not of type SIGNAGE """
     objects = BaseInfrastructure.get_manager_cls(InfrastructureGISManager)()
-    in_structure = InfrastructureStructureManager()
 
     class Meta:
-        proxy = True
+        db_table = 'a_t_infrastructure'
         verbose_name = _(u"Infrastructure")
         verbose_name_plural = _(u"Infrastructures")
 
@@ -195,19 +187,12 @@ class SignageGISManager(gismodels.GeoManager):
         return all_years
 
 
-class SignageStructureManager(StructureRelatedManager):
-    """ Overide default structure related manager, and filter by type. """
-    def get_queryset(self):
-        return super(SignageStructureManager, self).get_queryset().filter(type__type=INFRASTRUCTURE_TYPES.SIGNAGE)
-
-
-class Signage(BaseInfrastructure):
+class Signage(MapEntityMixin, BaseInfrastructure):
     """ An infrastructure in the park, which is of type SIGNAGE """
     objects = BaseInfrastructure.get_manager_cls(SignageGISManager)()
-    in_structure = SignageStructureManager()
 
     class Meta:
-        proxy = True
+        db_table = 'a_t_signaletique'
         verbose_name = _(u"Signage")
         verbose_name_plural = _(u"Signages")